inclusão metodo findLivroById no back

+++ livraria/src/main/java/br/senai/sp/livraria/dto/LivroTabelaDTO.java
@@ -46,7 +46,7 @@ public class LivroTabelaDTO {
 
 	private String imagem;
 
-	private BigDecimal preco;
+	private float preco;
 
 	private TipoLivro tipoLivro;
